# Reduce the number of allocations on the LevenshteinDistance function