@lithammer 2022-05-03
Released v1.1.5 with this.